re structure proposalDetailsPage folder
