+++ app/components/views/ProposalDetailsPage/ProposalDetails.module.css
