+++ app/components/views/ProposalDetailsPage/helpers/ProposalAbandoned.jsx
@@ -1,0 +1,7 @@
+import styles from "../ProposalDetails.module.css";
+import { FormattedMessage as T } from "react-intl";
+
+const ProposalAbandoned = () =>
+  <div className={styles.notVoting}><T id="proposalDetails.votingInfo.abandoned" m="Proposal has been abandoned" /></div>;
+
+export default ProposalAbandoned;
