+++ app/components/views/ProposalDetailsPage/helpers/index.js
@@ -1,0 +1,14 @@
+export { default as OverviewField } from "./OverviewField";
+export { default as NoElligibleTicketsVotingInfo } from "./NoElligibleTicketsVotingInfo";
+export { default as NoTicketsVotingInfo } from "./NoTicketsVotingInfo";
+export { default as ProposalVoted } from "./ProposalVoted";
+export { default as ProposalNotVoting } from "./ProposalNotVoting";
+export { default as ProposalAbandoned } from "./ProposalAbandoned";
+export { default as ProposalError } from "./ProposalError";
+export { default as TimeValue } from "./TimeValue";
+export { default as ProposalText } from "./ProposalText";
+export { default as Header } from "./Header";
+export { default as VoteInfo } from "./VoteInfo";
+export { default as OverviewVotingProgressInfo } from "./OverviewVotingProgressInfo";
+export { default as ChooseOptions } from "./ChooseOptions";
+export { default as ChooseVoteOption } from "./ChooseVoteOption";
