+++ app/components/views/ProposalDetailsPage/hooks.js
