+++ app/containers/Wallet.js
@@ -14,7 +14,7 @@ import TransactionPage from "components/views/TransactionPage";
 import TicketsPage from "components/views/TicketsPage";
 import TutorialsPage from "components/views/TutorialsPage";
 import GovernancePage from "components/views/GovernancePage";
-import ProposalDetailsPage from "components/views/ProposalDetails";
+import ProposalDetailsPage from "components/views/ProposalDetailsPage/ProposalDetailsPage";
 import TrezorPage from "components/views/TrezorPage";
 import LNPage from "components/views/LNPage";
 import SideBar from "components/SideBar";
